+++ tests/rule_tests/mod.rs
@@ -138,6 +138,7 @@ mod inject_value;
 mod no_local_function;
 mod remove_call_parens;
 mod remove_comments;
+mod remove_compound_assignment;
 mod remove_empty_do;
 mod remove_method_definition;
 mod remove_nil_declaration;